Move the auto-generate and clear feature from section into part of the special section.
