+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Extends/Notes/NoteExtend.cs
@@ -16,8 +16,6 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Notes
 
         private readonly IBindable<NoteEditMode> bindableMode = new Bindable<NoteEditMode>();
 
-        private readonly IBindable<NoteEditModeSpecialAction> bindableModeSpecialAction = new Bindable<NoteEditModeSpecialAction>();
-
         public NoteExtend()
         {
             bindableMode.BindValueChanged(e =>
@@ -31,7 +29,6 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Notes
                             new NoteConfigSection(),
                             new NoteSwitchSpecialActionSection(),
                         };
-                        updateActionArea();
                         break;
 
                     case NoteEditMode.Edit:
@@ -55,44 +52,12 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Notes
                         return;
                 }
             }, true);
-
-            bindableModeSpecialAction.BindValueChanged(e =>
-            {
-                if (bindableMode.Value != NoteEditMode.Generate)
-                    return;
-
-                updateActionArea();
-            }, true);
-
-            void updateActionArea()
-            {
-                RemoveAll(x => x is NoteAutoGenerateSection or NoteClearSection);
-
-                switch (bindableModeSpecialAction.Value)
-                {
-                    case NoteEditModeSpecialAction.AutoGenerate:
-                        Add(new NoteAutoGenerateSection());
-                        break;
-
-                    case NoteEditModeSpecialAction.SyncTime:
-                        // todo: implement
-                        break;
-
-                    case NoteEditModeSpecialAction.Clear:
-                        Add(new NoteClearSection());
-                        break;
-
-                    default:
-                        return;
-                }
-            }
         }
 
         [BackgroundDependencyLoader]
         private void load(IEditNoteModeState editNoteModeState)
         {
             bindableMode.BindTo(editNoteModeState.BindableEditMode);
-            bindableModeSpecialAction.BindTo(editNoteModeState.BindableSpecialAction);
         }
     }
 }
